cleaned up routes and server
